requirements.txt
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
blinker==1.3
certifi==0.0.8
chardet==2.2.1
cssselect==0.6.1
pyquery==1.2.4
itsdangerous==0.23
python-dateutil==2.2
requests==2.1.0
simplejson==3.3.1
six==1.4.1
Twisted==13.2.0
psycopg2==2.5
APScheduler==2.0.3
gevent==1.0
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
greenlet==0.4.1
Flask==0.8
Flask-SQLAlchemy==1.0
praw==2.1.20
gunicorn==0.14.3
Flask-Sockets==0.1

